Do not update Symbol defTrees when retyping after Inlining
The scaladoc for the defTree method in Symbol states: "The tree defining
the symbol at pickler time ...", but that was never really true.
Previously, since the setDefTree method was used in Typer, in any
retyping procedure those methods would be called again, with the main
example being erasure, where we would lose parts of type information from
those trees. Usually this was not an issue, since they weren't used after
Erasure.
However, the suspend mechanism used when compiling
macros with calls to them can cause the macro tree to go through
erasure, have their defTrees updated there, and then used for earlier
phases, with the problematic phase here being the init-checker, which
uses defTree calls extensively.
In the issue minimisation, init-checker was expecting a MethodType (due to the
missing TypeApply there), and instead got a PolyType, causing a crash.

Coincidentally in the past we would sometimes also get issues about
.tree method in Quotes returning unexpected stuff due to the same issue,
so this should also help there.